incorporating Kirill's code
